Have L1bProj in place and working. Skeleton of L1bTpCollect, but this needs work.

+++ python/lib/l1b_correct.py
@@ -1,5 +1,6 @@
 import geocal
 import logging
+import os
 
 logger = logging.getLogger('l1b_geo_process.l1b_geo_generate')
 
@@ -11,11 +12,11 @@ class L1bCorrect:
         self.igccol_initial = igccol_initial
         self.l1b_geo_config = l1b_geo_config
         self.geo_qa = geo_qa
+        if(not os.path.exists("extra_python_init.py")):
+            with open("extra_python_init.py", "w") as fh:
+                print("from emit import *\n", file=fh)
 
     def run(self):
-        # Save file so downstream programs know to import emit
-        with open("extra_python_init.py", "w") as fh:
-            print("from emit import *\n", file=fh)
         geocal.write_shelve("igccol_initial.xml", self.igccol_initial)
             
         self.igccol_corrected = self.igccol_initial
